Moving NoticeDescriptor builder to top-level class
