+++ src/main/java/org/folio/circulation/domain/notice/NoticeDescriptor.java
@@ -60,56 +60,4 @@ public class NoticeDescriptor {
     return recurringPeriod;
   }
 
-
-  public static class Builder {
-
-    private String templateId;
-    private NoticeFormat noticeFormat;
-    private NoticeEventType noticeEventType;
-    private NoticeTiming timing;
-    private Period timingPeriod;
-    private boolean recurring;
-    private Period recurringPeriod;
-
-    public Builder setTemplateId(String templateId) {
-      this.templateId = templateId;
-      return this;
-    }
-
-    public Builder setNoticeFormat(NoticeFormat noticeFormat) {
-      this.noticeFormat = noticeFormat;
-      return this;
-    }
-
-    public Builder setNoticeEventType(NoticeEventType noticeEventType) {
-      this.noticeEventType = noticeEventType;
-      return this;
-    }
-
-    public Builder setTiming(NoticeTiming timing) {
-      this.timing = timing;
-      return this;
-    }
-
-    public Builder setTimingPeriod(Period timingPeriod) {
-      this.timingPeriod = timingPeriod;
-      return this;
-    }
-
-    public Builder setRecurring(boolean recurring) {
-      this.recurring = recurring;
-      return this;
-    }
-
-    public Builder setRecurringPeriod(Period recurringPeriod) {
-      this.recurringPeriod = recurringPeriod;
-      return this;
-    }
-
-    public NoticeDescriptor build() {
-      return new NoticeDescriptor(
-        templateId, noticeFormat, noticeEventType,
-        timing, timingPeriod, recurring, recurringPeriod);
-    }
-  }
 }
